refactor(category-sync): rename file
