+++ lib/sync/category-actions.js
@@ -1,5 +1,5 @@
 import forEach from 'lodash.foreach'
-import { buildBaseAttributesAction } from './utils/actions-utils'
+import { buildBaseAttributesAction } from './utils/common-actions'
 import * as diffpatcher from './utils/diffpatcher'
 
 function actionsBaseList () {
