+++ lib/sync/product-actions.js
@@ -3,7 +3,7 @@ import forEach from 'lodash.foreach'
 import unique from 'lodash.uniq'
 import { find } from './utils/array'
 import * as diffpatcher from './utils/diffpatcher'
-import { buildBaseAttributesAction } from './utils/actions-utils'
+import { buildBaseAttributesAction } from './utils/common-actions'
 
 const REGEX_NUMBER = new RegExp(/^\d+$/)
 const REGEX_UNDERSCORE_NUMBER = new RegExp(/^\_\d+$/)
