+++ lib/sync/utils/common-actions.js
